Add and improve test cases for FileUtil

+++ src/main/java/com/khanh/sample/utils/FileUtil.java
@@ -14,11 +14,8 @@ public class FileUtil {
     public static List<File> getNewFiles(String directoryPath, long lastCheck) throws IOException {
         List<File> newFiles = new ArrayList<>();
 
-        File[] files = getFiles(directoryPath, LastModifiedFileComparator.LASTMODIFIED_REVERSE);
-        for (File file : files) {
-            if(file.isDirectory()) {
-                continue;
-            }
+        List<File> allFiles = getFiles(directoryPath, LastModifiedFileComparator.LASTMODIFIED_REVERSE);
+        for (File file : allFiles) {
             if (file.lastModified() > lastCheck) {
                 newFiles.add(file);
             } else {
@@ -29,15 +26,22 @@ public class FileUtil {
         return newFiles;
     }
 
-    public static File[] getFiles(String directoryPath, Comparator<File> comparator) throws IOException {
+    public static List<File> getFiles(String directoryPath, Comparator<File> comparator) throws IOException {
         File directory = new File(directoryPath);
         if (directory.isFile())
             throw new IOException("This path is not a directory.");
 
         File[] files = directory.listFiles();
-        Arrays.sort(files, comparator);
+        List<File> result = new ArrayList<>();
 
-        return files;
+        for(File file : files) {
+            if(file.isFile()) {
+                result.add(file);
+            }
+        }
+
+        Collections.sort(result, comparator);
+        return result;
     }
 
     public static void checkAndCreateDirectory(String filePath) {
@@ -51,27 +55,50 @@ public class FileUtil {
 
     public static boolean deleteDirectory(String path) {
         File directory = new File(path);
+        return deleteDirectory(directory);
+    }
+
+    public static boolean deleteDirectory(File directory) {
         if (directory.exists()) {
-            File[] files = directory.listFiles();
-            if (files != null) {
-                for (File file : files) {
-                    file.delete();
-                }
+            if(deleteFilesInDirectory(directory)) {
+                return directory.delete();
             }
-            return directory.delete();
+            return false;
         }
         return true;
     }
 
-    public static boolean copyFile(File source, File destination) throws IOException {
-        if(!destination.exists())
-            destination.mkdirs();
+    public static boolean deleteFilesInDirectory(File directory) {
+        File[] files = directory.listFiles();
 
-        if(destination.isDirectory()) {
-            String path = destination.getPath() + File.separator + source.getName();
-            Files.copy(source.toPath(), Path.of(path), StandardCopyOption.REPLACE_EXISTING);
-        } else {
-            Files.copy(source.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
+        if (files != null) {
+            int amount = files.length, count = 0;
+
+            // In order to delete folder, all files in the folder need to be deleted first
+            for (File file : files) {
+                if(file.isDirectory()) {
+                    if(deleteDirectory(file)) {
+                        count++;
+                    }
+                }
+                else {
+                    if (file.delete()) {
+                        count++;
+                    }
+                }
+            }
+            return amount == count;
+        }
+        return true;
+    }
+
+    public static boolean copyFile(File sourceFile, File destinationDirectory) throws IOException {
+        if(!destinationDirectory.exists())
+            destinationDirectory.mkdirs();
+
+        if(destinationDirectory.isDirectory()) {
+            String path = destinationDirectory.getPath() + File.separator + sourceFile.getName();
+            Files.copy(sourceFile.toPath(), Path.of(path), StandardCopyOption.REPLACE_EXISTING);
         }
         return true;
     }
